lesson-04/flask-application/requirements.txt
flask==3.1.0
requests==2.32.3
jsonify==0.5
psycopg2-binary==2.9.10
